refactor: proto, grpc service
