+++ .devcontainer/Dockerfile
@@ -14,6 +14,6 @@ RUN sudo apt-get update && \
     # install protobuf compiler tools
     # https://grpc.io/docs/languages/go/quickstart/#prerequisites
     sudo apt-get install -y protobuf-compiler && \
-    go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.28 && \
-    go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.2 && \
+    go install google.golang.org/protobuf/cmd/protoc-gen-go@latest && \
+    go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@latest && \
     export PATH="$PATH:$(go env GOPATH)/bin"
