+++ .devcontainer/install-dev-tools.sh
@@ -1,0 +1,2 @@
+# compile proto
+sh build/scripts/proto/compile.sh
