+++ examples/service/config.jsonnet
