+++ proto/config.proto
@@ -1,0 +1,11 @@
+syntax = "proto3";
+
+package substation.proto;
+
+option go_package = "github.com/brexhq/substation/proto";
+
+// Capsule mirrors the Capsule struct defined in config
+message Capsule {
+  bytes data = 1;
+  bytes metadata = 2;
+}
