--- substation.proto
@@ -1,21 +1,0 @@
-syntax = "proto3";
-
-option go_package = "github.com/brexhq/substation";
-
-package substation;
-
-service Substation {
-  // Sink mirrors the Sink interface defined in internal/sink
-  rpc Sink(stream Capsule) returns (Ack) {}
-}
-
-// Capsule mirrors the Capsule struct defined in config
-message Capsule {
-  bytes data = 1;
-  bytes metadata = 2;
-}
-
-// Ack is a server-side response that acknowledges the receipt of Capsules (or any streaming request)
-message Ack {
-  uint32 count = 1;
-}
